Push the just-shortened URL into the browser's history
This improves the user experience by presenting the shortened URL in the
browser's location bar.  This means that it can be pulled to the
clipboard *really quickly* by issuing Ctrl-L then Ctrl-C, ready for
posting to Jabber/IRC/4chan.

We use a tiny bit of HTML5 Javascript to do it.  It's arguably better
than some janky JS or Flash trying to screw with the user's clipboard.